Added explanations in Javadoc style and in-line comments. Added TODO: comments on how to fix issues
No code was changed, only comments added to aid in understanding the code base. Indenting altered to fit standaard.
TODO's were added. The best order to fix is start with the ones in Swerve_Module. Once those are done, do line 121 in Drive. after that finish the others in drive regarding SwerveDriveOdometry.